Added timetable function for db and app version api endpoint
